highlight current user chapter reaction selection
