+++ frontend/app/components/h5p-list-item.hbs
@@ -14,7 +14,7 @@
     <div class="card-body chapter-body">
       <div class="actions-section d-flex justify-content-between">
         <div class="reactions">
-          <Reaction @model={{@chapter}} @reactAction={{@userChapterReactionAction}} />
+          <Reaction @chapter={{@chapter}}/>
         </div>
         <div class="popup-actions">
           <button class="btn-popup-actions">
