+++ frontend/app/components/h5p-list-item.js
@@ -1,9 +1,8 @@
 import Component from '@glimmer/component';
-import { inject } from '@ember/service';
+import {inject} from '@ember/service';
 
 export default class H5pListItemComponent extends Component {
-  @inject
-  config
+  @inject config;
 
   color = Math.floor(Math.random() * 16777215).toString(16);
