+++ frontend/app/styles/_chapters.scss
@@ -1,5 +1,5 @@
 $palette: (
-  primary: $primary,
+  primary: #F57010,
   white: #FFFFFF,
   gray30: #3F484D,
   gray45: rgb(94, 108, 115),
