+++ frontend/app/styles/bs4.scss
@@ -10,7 +10,7 @@ $indigo: #6610f2;
 $purple: #670098;
 $pink: #e83e8c;
 $red: #dc3545;
-$orange: #FF6700;
+$orange: #ff6700;
 $yellow: #FFD300;
 $green: #00aa83;
 $teal: #FFEBAF;
