+++ frontend/app/templates/home.hbs
@@ -4,9 +4,7 @@
   <div class="row mt-5 ">
     {{#each model as |chapter|}}
 
-      <H5pListItem @chapter={{chapter}}
-                   @userChapterReactionAction={{this.userChapterReaction}}
-                   @reactionNotify={{this.reactionFeedback}}/>
+      <H5pListItem @chapter={{chapter}} @reactionNotify={{this.reactionFeedback}}/>
     {{/each}}
   </div>
 </div>
